Cert manager

+++ test/helpers/wallet.spec.ts
@@ -19,30 +19,67 @@ import { FileSystemWallet } from 'fabric-network';
 import * as fs from 'fs';
 import * as rimraf from 'rimraf';
 
+// tslint:disable-next-line: no-var-requires
+const CertificateManagerWallet = require('@blockchainlabs/ibm-certificate-manager-wallet');
 import * as walletHelper from '../../server/helpers/wallet';
 
-const cert = fs.readFileSync(`${__dirname}/../mocks/cert`, 'utf8');
-const key = fs.readFileSync(`${__dirname}/../mocks/key`, 'utf8');
+const cert = fs.readFileSync(`${__dirname}/../mocks/testuser1.pem`, 'utf8');
+const key = fs.readFileSync(`${__dirname}/../mocks/testuser1.key`, 'utf8');
+
+const FILESYSTEM_WALLET = 'FileSystemWallet';
+const CERTIFICATE_MANAGER_WALLET = 'CertificateManagerWallet';
 
 describe('helpers - wallet', () => {
-    beforeAll(() => {
-        // delete test wallet before starting
-        rimraf.sync(config.get('fsWalletPath'));
+
+    beforeAll(async () => {
+        // check if correct wallet type in config
+        const supportedWallets: string[] = config.get('supportedWallets');
+        if (!(supportedWallets.includes(config.get('activeWallet')))) {
+            throw new Error ('Incorrect activeWallet in config');
+        }
+
+        // delete test wallet before starting for file system wallet
+        if (config.get('activeWallet') === FILESYSTEM_WALLET) {
+            rimraf.sync(config.get('fsWalletPath'));
+        }
     });
-    afterAll(() => {
-        // delete test wallet after tests
-        rimraf.sync(config.get('fsWalletPath'));
+    afterAll(async () => {
+        // delete test wallet after tests for file system wallet
+        if (config.get('activeWallet') === FILESYSTEM_WALLET) {
+            rimraf.sync(config.get('fsWalletPath'));
+        }
+
+        // delete test user after tests for certifacte manager wallet
+        if (config.get('activeWallet') === CERTIFICATE_MANAGER_WALLET) {
+            if (await walletHelper.identityExists('testuser1')) {
+                await walletHelper.deleteIdentity('testuser1');
+            }
+        }
     });
 
     let wallet;
 
     beforeEach(() => {
+        walletHelper.initWallet(config.get('activeWallet'));
         wallet = walletHelper.getWallet();
     });
 
+    describe('#initWallet', () => {
+        test('should fail for incorrect type of wallet', () => {
+            expect(() => {
+                walletHelper.initWallet('IncorrectWalletType');
+            }).toThrow(new Error('Incorrect activeWallet in config'));
+        });
+    });
+
     describe('#getWallet', () => {
-        test('should get wallet of type FileSystemWallet', () => {
-            expect(wallet).toBeInstanceOf(FileSystemWallet);
+        test('should get correct type of wallet', () => {
+            if (config.get('activeWallet') === FILESYSTEM_WALLET) {
+                expect(wallet).toBeInstanceOf(FileSystemWallet);
+            }
+            if (config.get('activeWallet') === CERTIFICATE_MANAGER_WALLET) {
+                expect(wallet).toBeInstanceOf(CertificateManagerWallet);
+            }
         });
     });
 
@@ -65,6 +102,29 @@ describe('helpers - wallet', () => {
         test('should return false if identity does not exist in wallet', async () => {
             const result = await walletHelper.identityExists('testuser2');
             expect(result).toBe(false);
+        });
+    });
+
+    describe('#deleteIdentity', () => {
+        test('should handle error during delete', async () => {
+            // let wallet;
+            (wallet.delete) = jest.fn(() => {
+                return Promise.reject(new Error('error deleting identity'));
+            });
+            try {
+                await walletHelper.deleteIdentity('testuser1');
+            } catch (e) {
+                expect(e.message).toMatch('error deleting identity');
+            }
+        });
+
+        test('should delete identity properly', async () => {
+            expect(await walletHelper.importIdentity('testuser2', 'org1', cert, key));
+            const result = await walletHelper.identityExists('testuser2');
+            expect(result).toBe(true);
+            await walletHelper.deleteIdentity('testuser2');
+            const resultFalse = await walletHelper.identityExists('testuser2');
+            expect(resultFalse).toBe(false);
         });
     });
 });
